Reorganisation of openstack package
